+++ NEWS.rst
@@ -14,6 +14,8 @@ Version History
 * Introduce a bench.py script for performing (crude) benchmarks.
 * ZSTD_CCtx instances are now reused across multiple compress() operations.
 * ZstdCompressor.write_to() now has a flush() method.
+* ZstdCompressor.compressobj()'s flush() method now accepts an argument to
+  flush a block (as opposed to ending the stream).
 
 0.5.2 (released 2016-11-12)
 ---------------------------